Fix the build with -fno-common
These two globals were defined in a header, so they "appeared" in multiple
objects. The previous compiler default, -fcommon, would put them into a
common section and merge all of the definitions. In LLVM11 and GCC10, the
default switch to -fno-common and made this an error.

Extern them and plop their definitions in some .c files that seem suitable.